requirements.txt
fastapi==0.109.1
uvicorn==0.23.1
mysql-connector-python==8.0.25
python-dotenv==1.0.0
